Only users with an open battle request will be planned in

+++ app/Jobs/PlanBattles.php
@@ -22,10 +22,20 @@ class PlanBattles implements ShouldQueue
     public function handle(): void
     {
         foreach (BattleRequest::all() as $battle_request) {
+            if (! BattleRequest::query()->where('id', $battle_request->id)->exists()) {
+                return;
+            }
+
             $battle = $this->planBattleFor($battle_request->user);
 
             if ($battle) {
                 $battle_request->delete();
+
+                if ($battle_request->user->id === $battle->hero->id) {
+                    $battle->monster->battleRequests()->first()->delete();
+                } else {
+                    $battle->hero->battleRequests()->first()->delete();
+                }
             }
         }
     }
